[Update] To search product function is modified. And to add product name when clicked is updated
